Add files via upload
In this project, I:

-Used Visualization to answer to some analytical questions
-Explored different graphs needed to illustrate insights in my dataset 
-Utilised Plotly module in python

In this project we are going to use COVID19 dataset we have consisting of data-related cumulative number of confirmed, recovered, and deaths cases. we are going to prepare this dataset to answer these questions: How does Global Spread of the virus look like? How intensive the spread of the virus has been in the countries? Does covid19 national lockdowns and self-isolations in different countries have actually impact on COVID19 transmission? we are going to use Plotly module to plot some insightful and intuitive graphs to answer the questions.